Move testhelpers.go contents to testhelpers_test.go
